Add Roblox server verification, make more modular and modified README
Added Roblox server verification to all forms of API requests, which requires now with each request:

- A game server's JobId

This prevents malicious requests from hitting the API - and requires an active game instance in order to validate.

Modularity is also introduced with this commit, which will help isolate indifidual features of the project into seperate files.

Modified README to be more concise,  added a Roblox account as a prerequisite.