Remove reliance on placeId, make locals modular, add auth.js
This is a slight overhaul of the backend code, to not rely on specific places (to remove the need for configuring the script yourself if you want to host it yourself) and make local script use modules instead of a model, to allow people to use the script builder project with minimal configuration required on their part to "get up and go".

This is significant since people only need to require the module and customize their options in their own scripts requiring the provided model, rather than doing everything themselves with the original approach.

The customizability is still there for people who want to not use this method, however - and is relatively trivial to do even then.

This also updates a dependency which fixes CVE-2020-15168 (low severity) - a potential DoS vulunerability inside node-fetch. It doesn't really impact this project, but if it ever did - it would cost a lot of money if someone were to abuse it in computing resources.